fixed data types

+++ python_introduction/finance_calculator.py
@@ -1,5 +1,5 @@
-income = int(input("Enter your monthly income: "))
-expenses = int(input("Enter your total monthly expenses: "))
+income = float(input("Enter your monthly income: "))
+expenses = float(input("Enter your total monthly expenses: "))
 annual_interest = 0.05
 monthly_savings = income - expenses
 projected_savings = (monthly_savings * 12) + (monthly_savings*12*0.05)
